+++ examples/worlds/lrauv_control_demo.sdf
